[BPF] start ipv6 parsing
